# docs/openapi/taakhelden-core-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wispel core API",
    "version": "0.1.0",
    "description": "Generated core contract snapshot from packages/shared. Covers web dashboard and iOS Phase 1 endpoints."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a parent and family",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in a parent account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/apple": {
      "post": {
        "summary": "Sign in with Apple",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleAuthBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Rotate a parent refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Request a password-reset email (always 200; no email enumeration)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset password with a one-time token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/family-code": {
      "post": {
        "summary": "Resolve a family code to child profiles",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyCodeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyCodeResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/child-session": {
      "post": {
        "summary": "Pair a child device session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildSessionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/child-session/refresh": {
      "post": {
        "summary": "Rotate a child device refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChildSessionRefreshBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildSessionResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "$ref": "#/definitions/ApiError",
        "definitions": {
          "ApiError": {
            "type": "object",
            "properties": {
              "error": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "enum": [
                      "VALIDATION_FAILED",
                      "UNAUTHORIZED",
                      "FORBIDDEN",
                      "NOT_FOUND",
                      "TASK_ALREADY_COMPLETED",
                      "INVALID_STATUS",
                      "UNDO_WINDOW_EXPIRED",
                      "EMAIL_IN_USE",
                      "INVALID_CREDENTIALS",
                      "INSUFFICIENT_POINTS",
                      "REWARD_LIMIT_REACHED",
                      "PIN_LOCKED",
                      "INVALID_FAMILY_CODE",
                      "INVALID_INVITE",
                      "INSTANCE_SLOT_TAKEN",
                      "IDEMPOTENCY_REPLAY",
                      "IDEMPOTENCY_KEY_REUSED",
                      "RATE_LIMITED",
                      "UPSTREAM_UNAVAILABLE"
                    ]
                  },
                  "message": {
                    "type": "string"
                  },
                  "details": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                },
                "required": [
                  "code",
                  "message"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "error"
            ],
            "additionalProperties": false
          }
        }
      },
      "RegisterBody": {
        "$ref": "#/definitions/RegisterBody",
        "definitions": {
          "RegisterBody": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              },
              "password": {
                "type": "string",
                "minLength": 10
              },
              "familyName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 50
              },
              "displayName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 30
              },
              "turnstileToken": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "password",
              "familyName",
              "displayName",
              "turnstileToken"
            ],
            "additionalProperties": false
          }
        }
      },
      "LoginBody": {
        "$ref": "#/definitions/LoginBody",
        "definitions": {
          "LoginBody": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              },
              "password": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "password"
            ],
            "additionalProperties": false
          }
        }
      },
      "AppleAuthBody": {
        "$ref": "#/definitions/AppleAuthBody",
        "definitions": {
          "AppleAuthBody": {
            "type": "object",
            "properties": {
              "identityToken": {
                "type": "string",
                "minLength": 1
              },
              "familyName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 50
              },
              "displayName": {
                "type": "string",
                "minLength": 1,
                "maxLength": 30
              }
            },
            "required": [
              "identityToken"
            ],
            "additionalProperties": false
          }
        }
      },
      "FamilyCodeBody": {
        "$ref": "#/definitions/FamilyCodeBody",
        "definitions": {
          "FamilyCodeBody": {
            "type": "object",
            "properties": {
              "familyCode": {
                "type": "string",
                "minLength": 6,
                "maxLength": 6
              }
            },
            "required": [
              "familyCode"
            ],
            "additionalProperties": false
          }
        }
      },
      "ChildSessionBody": {
        "$ref": "#/definitions/ChildSessionBody",
        "definitions": {
          "ChildSessionBody": {
            "type": "object",
            "properties": {
              "familyCode": {
                "type": "string",
                "minLength": 6,
                "maxLength": 6
              },
              "childId": {
                "type": "string"
              },
              "pincode": {
                "type": "string",
                "pattern": "^\\d{4}$"
              }
            },
            "required": [
              "familyCode",
              "childId",
              "pincode"
            ],
            "additionalProperties": false
          }
        }
      },
      "ChildSessionRefreshBody": {
        "$ref": "#/definitions/ChildSessionRefreshBody",
        "definitions": {
          "ChildSessionRefreshBody": {
            "type": "object",
            "properties": {
              "refreshToken": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "refreshToken"
            ],
            "additionalProperties": false
          }
        }
      },
      "ChildSessionResult": {
        "$ref": "#/definitions/ChildSessionResult",
        "definitions": {
          "ChildSessionResult": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              },
              "expiresIn": {
                "type": "number"
              },
              "child": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "avatarId": {
                    "type": "string",
                    "nullable": true
                  },
                  "ageMode": {
                    "type": "string",
                    "enum": [
                      "young",
                      "mid",
                      "teen"
                    ]
                  }
                },
                "required": [
                  "id",
                  "displayName",
                  "avatarId",
                  "ageMode"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "accessToken",
              "refreshToken",
              "expiresIn",
              "child"
            ],
            "additionalProperties": false
          }
        }
      },
      "FamilyCodeResult": {
        "$ref": "#/definitions/FamilyCodeResult",
        "definitions": {
          "FamilyCodeResult": {
            "type": "object",
            "properties": {
              "familyName": {
                "type": "string"
              },
              "children": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "avatarId": {
                      "type": "string",
                      "nullable": true
                    },
                    "ageMode": {
                      "type": "string",
                      "enum": [
                        "young",
                        "mid",
                        "teen"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "displayName",
                    "avatarId",
                    "ageMode"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "familyName",
              "children"
            ],
            "additionalProperties": false
          }
        }
      },
      "ParentSessionResult": {
        "$ref": "#/definitions/ParentSessionResult",
        "definitions": {
          "ParentSessionResult": {
            "type": "object",
            "properties": {
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              },
              "expiresIn": {
                "type": "number"
              },
              "userId": {
                "type": "string"
              },
              "familyId": {
                "type": "string"
              }
            },
            "required": [
              "accessToken",
              "refreshToken",
              "expiresIn",
              "userId",
              "familyId"
            ],
            "additionalProperties": false
          }
        }
      },
      "RefreshBody": {
        "$ref": "#/definitions/RefreshBody",
        "definitions": {
          "RefreshBody": {
            "type": "object",
            "properties": {
              "refreshToken": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "refreshToken"
            ],
            "additionalProperties": false
          }
        }
      },
...
      "ForgotPasswordBody": {
        "$ref": "#/definitions/ForgotPasswordBody",
        "definitions": {
          "ForgotPasswordBody": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              }
            },
            "required": [
              "email"
            ],
            "additionalProperties": false
          }
        }
      },
      "ResetPasswordBody": {
        "$ref": "#/definitions/ResetPasswordBody",
        "definitions": {
          "ResetPasswordBody": {
            "type": "object",
            "properties": {
              "token": {
                "type": "string",
                "minLength": 1
              },
              "password": {
                "type": "string",
                "minLength": 10
              }
            },
            "required": [
              "token",
              "password"
            ],
            "additionalProperties": false
          }
        }
      }
    }
  }
}